requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
pyyaml>=6.0
joblib>=1.3
matplotlib>=3.7